+++ tests/testthat/test-reporter-debug.R
@@ -1,4 +1,5 @@
 test_that("produces consistent output", {
+  local_edition(2)
   local_mock(
     show_menu = function(choices, title = NULL) {
       cat(paste0(format(seq_along(choices)), ": ", choices, sep = "\n"), "\n", sep = "")
@@ -16,6 +17,7 @@ get_vars_from_debug_reporter <- function(choice, fun, envir = parent.frame()) {
 }
 
 get_frame_from_debug_reporter <- function(choice, fun, envir = parent.frame()) {
+  local_edition(2)
   force(choice)
   test_debug_reporter_parent_frame <- NULL